net: don't enclose non-literal IPv6 addresses in square brackets
RFC 4007 describes guidelines for the use of textual representation of
IPv6 addressing/routing scope zone and doesn't prohibit to use the
format for implementation dependent purposes. However, a non-literal
IPv6 address, for example, a host name, with a zone enclosed in square
brackets confuses us because a zone is basically for non-global
addresses and a pair of square brackets is used as a set of delimiters
between a litearal IPv6 address and a service name or transport port
number.

To avoid such confusion, this changes makes JoinHostPort not do
enclose non-literal IPv6 addresses in square brackets to recommend that
anything enclosed in square brackets should be a literal IPv6 address.

Also updates docs and test cases on SplitHostPort and JoinHostPort for
clarification.

Fixes #18059.
Fixes #18060.

Change-Id: I5c3ccce4fa0fbdd58f698fc280635ea4a14d2a37