Update patch set 3
Patch Set 3:

My main concern with the test is that if it fails, can I quickly figure out which input it failed on?  I'd print the test index, the value index, and "idx" so that it would be clear where things left the rails.

Patch-set: 3
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>