Update patch set 3
Patch Set 3:

> My main concern with the test is that if it fails, can I quickly
 > figure out which input it failed on?  I'd print the test index, the
 > value index, and "idx" so that it would be clear where things left
 > the rails.

Yeah, I had similar concerns. I'll add some more information so it is easier to isolate the failing test programmatically (at the moment the easiest way is to grep for the failing comparison).

Another concern I have is that it is hard to use GOSSAFUNC to debug the failing function because it will be anonymous. Naming all the functions makes the test file a lot bigger though and this isn't likely to be a test that fails a lot.

Patch-set: 3
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>