Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > .. Yes, it will break if application under wine will change the
>  > time - time in golang will not be changed.
> 
> That is the problem. Go app under Wine will not notice that time got changed in the OS. Are you going to address that problem?

The PROBLEM is that go app under Wine just does.not.work at all now.

Please remember that currently it is virtually impossible to start application under wine which uses anything but time.Now() and even this is broken. It does not work at all because of limited windows support by wine which is implicitly required by golang.

I propose a fix which works in every cases but time change in application itself. Just few months back golang didn't have monotonic time at all and changing time virtually broke every application which ended up with huge number of internet posts, this patch makes no worse than that, and in fact allows to work almost all apps, in fact, I do not know any application which changes time internally if it is not a time daemon like NTP, which is useless under wine since wine is single-application system.

> I was suggesting you replace your _KSYSTEM_TIME reading code with call to GetSystemTime (or whatever function you like). This should solve the problem I am complaining about. What do you think?

What is the purpose? I would rather add another asm wrapper which calls memory reading stuff and exports it into runtime package so it could be used instead of reimplementing it manually in golang, but I already did that and that part of the patch was asked to be thrown out.

Patch-set: 9
Reviewer: Gerrit User 17055 <17055@62eb7196-b449-3ce5-99f1-c037f21e1705>