Update patch set 1
Patch Set 1:

(1 comment)

> Code looks good.  I think we want to check this in to master and
> then backport it, though.

CL 40433 is in master and this is the cherry-pick of that CL onto the release branch. AFAICT it is not usual to mention that this is a cherry-pick in the commit message - but maybe that would make sense. I should probably have added a comment to Gerrit at least.

> The process is still a bit in flux.  See
> the golang-dev thread and https://github.com/golang/go/issues/19776

Yeah, I was trying to manually go through the automatic process outlined in https://golang.org/issues/19776.

 > whomever does the point release will do 4-6.

I've been doing the cherry picks for s390x specific issues recently. I feel a bit better about doing them myself since it makes them less likely to get overlooked and I can fix any conflicts that arise. Maybe it would be better for the person cutting the release to handle all the cherry-picks, I'm not sure.

Patch-set: 1
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>