Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Any particular reason for this?  These are C comments too as of C99, which was 18 years ago.

Just for maximum portability. As far as I can tell, this is the only place where C++ style comments are used.
By this change, we can compile code with -std=c89.
So I think this is worthy as a one-line commit.
If this change is big, I didn't send it.

Patch-set: 1
Reviewer: Gerrit User 11631 <11631@62eb7196-b449-3ce5-99f1-c037f21e1705>