cmd/internal/obj: ARM, use immediates instead of constant pool entries
When a constant doesn't fit in a single instruction, use two
paired instructions instead of the constant pool.  For example

  ADD $0xaa00bb, R0, R1

Used to rewrite to

  MOV ?(IP), R11
  ADD R11, R0, R1

Instead, do

  ADD $0xaa0000, R0, R1
  ADD $0xbb, R1, R1

Same number of instructions.
Good:
  4 less bytes (no constant pool entry)
  One less load.
Bad:
  Critical path is one instruction longer.

It's probably worth it to avoid the loads, they are expensive.

DO NOT SUBMIT

Alternative idea for CL 39552.  Fix in assembler instead of compiler.

Change-Id: Ib206836161fdc94a3962db6f9caa635c87d57cf1