Update patch set 1
Patch Set 1:

We should probably not do this. This is also a good reason why we never want to do this, even if the compiler's invariants would allow it w/o a need for testing for "rootedness". That in itself is worth a good comment so we don't make this mistake down the road.

Patch-set: 1
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>