Update patch set 12
Patch Set 12:

I'm not completely sure what the race is, but:

So I think the data race seems to be due to there not being a 'happens before' relation between writing to the labelMap when it's being created (on lines 50 and 53 in label.go) and reading from the labelMap when we're about to write to the proto (on line 349 in proto.go).

I've added dummy loads and stores to silence the race detector but I'm aware that this isn't the correct solution.

Patch-set: 12
Reviewer: Gerrit User 10033 <10033@62eb7196-b449-3ce5-99f1-c037f21e1705>