cmd/compile: add global flag Types.SizeCalculationDisabled
Type.Size and Type.Alignment are used differently
by the front and back ends.

For the front end, they calculate size and alignment if needed.
For the back end, they retrieve existing calculations,
and must fail if they are not already calculated.
Enforce that.

This is an alternative to CL 42016.

Change-Id: I8c7b4410ee4c2a68573102f6b9b635f4fdcf392e