Update patch set 6
Patch Set 6:

> Harshavardhana, I do not even know what "NTFS volumes with
 > Deduplication enabled" is, but I suspect https://go-review.googlesource.com/#/c/41834/
 > will fix issue #10935. You say you have some "NTFS volumes with
 > Deduplication enabled" setup. Can you verify that issue #10935 is
 > fixed by CL 41834. Thank you.
 > 
 > Alex

Yes i already tested it does not fix the issue in #10935. An additional check is needed in Mode() which is performed in this PR.

NTFS volumes have a de-duplication feature, from what it does it looks like a file level de-duplication. So the files which have been de-duplicated have file attributes FILE_ATTRIBUTE_REPARSE_POINT but current code treats all files with this FILE_ATTRIBUTE_REPARSE_POINT attribute as symlink which is not the case. 

FILE_ATTRIBUTE_REPARSE_POINT along with ReparseTag of IO_REPARSE_TAG_SYMLINK is a symlink. Additionally IO_REPARSE_TAG_MOUNT_POINT is also a symlink. 

ReparseTag for files which have been de-duped will have a ReparseTag of IO_REPARSE_TAG_DEDUP.

The reason i cannot add tests is because you would need an NTFS volume with the above mentioned feature enabled and also to have a file which has been de-duped. There seems to be no way to use CreateFile() Win32 API to do this. 

I would love to get tests for this in this PR but i kind of wasted an entire day figuring that out :( . Some guidance is helpful on how to proceed here.

Patch-set: 6
Reviewer: Gerrit User 11900 <11900@62eb7196-b449-3ce5-99f1-c037f21e1705>