Update patch set 7
Patch Set 7:

Also what I think would be helpful (since we cannot have test here), is to at least steps to reproduce this problem documented here in CL description or in the issue. Something that outside person can follow to see what the issue is, and whether it is fixed by this CL or not.

Alex

Patch-set: 7