Update patch set 2
Patch Set 2:

> Thanks for the speedy review. I'd also like Matthew to look at this; these are dangerous waters.

Feel free to add checkwidth calls in the frontend (;)) as necessary. That should be a completely safe operation, other than maybe it means we end up calculating the width of more types than really necessary.

That said, hopefully we can figure out a principled approach to it. It certainly doesn't feel robust to scatter them around.

Patch-set: 2