Update patch set 3
Patch Set 3:

So i tested this looks like this PR solves the problem which is being solved by CL #41504. 

The fix was perhaps simple as avoiding Readlink() i.e which uses CreateFile() but uses REPARSE_POINT attribute to look at the file. So the check in Mode() now is ensured that every REPARSE_POINT is symlink. 

But i would suggest that we use the Reserved0 filed in FileWithData structure and validate it properly in Mode(). 

Thanks for the fix Alex. Will abandon my change in CL #41504.

Patch-set: 3
CC: Gerrit User 11900 <11900@62eb7196-b449-3ce5-99f1-c037f21e1705>