Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)

si is the index for src. I need it in decodeQuantum so the error index are correct. Maybe there is another way to do this? 
I'm away from my computer for the next week and if I'm correct, we are reaching the end of the merge period for new code so barring any other comment, if we decide it's a matter of documenting si, we can merge this patch and I'll improve the comment in another patch when I'm back. Otherwise we shouldn't rush it and this can wait until it's up to our standards of quality.

Patch-set: 5
Reviewer: Gerrit User 16720 <16720@62eb7196-b449-3ce5-99f1-c037f21e1705>