Update patch set 3
Patch Set 3:

I'm not sure what the tradeoff is here.  You're getting rid of some Progs, but then generating a bunch of static data.
That static data has a bunch of relocations in it.
But then maybe the explicit call list also has relocations for the call and for the stackmap?  Or is the stackmap just an int?

In any case, I'm confused as to the mechanism by which this CL is a net improvement.  It isn't as simple as just reducing Prog usage.

Patch-set: 3
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>