cmd/dist: spin up fewer test dir runners
5 shards, each of which spins up NumCPU processes,
each of which is running at GOMAXPROCS=NumCPU,
is too much. It makes my laptop unusable.

It might also be in part responsible for test flakes
that require a moderately responsive system,
like #18589 (backedge scheduling) and #19276 (locklinear).

It's possible that Go should be a better neighbor in general;
that's #17969. In the meantime, fix this corner of the world.

While we're here, give the test dir test a better name,
so that it is more usable with 'go tool dist test -run';
almost all tests match the regexp "test".

This is a partial reversion of CL 18199.
It leaves the sharding infrastructure intact for
future use and experimentation, and to make
this CL lower risk for the freeze.

Fixes #20141.

Change-Id: I123cf9436f4f4da3550372896265c38117b78071