crypto/elliptic: fix incomplete addition used in CombinedMult.
DO NOT SUBMIT: I need to ask the original author of the P-256 code to review this first.

The optimised P-256 includes a CombinedMult function, which doesn't do
dual-scalar multiplication, but does avoid an affine conversion for
ECDSA verification.

However, it currently uses an assembly point addition function that
doesn't handle exceptional cases.

What remains to be done:
   1. Ask Vlad to check this over.
   2. S390x copied this bug and so that code will also need to be
      updated.

Fixes #20215.

Change-Id: I4ba2ca1a546d883364a9bb6bf0bdbc7f7b44c94a