Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> As mentioned in https://go-review.googlesource.com/c/42501/ I think we should not encourage people to use broken cryptography even if they may use it for non-security purposes.
> But if it's important to explicitly outline the non-security part, I'm okay with that.

The original text implied the package itself was broken, not the algorithm it implements. With that reading, what was one to do if one needed DES, say, to talk to some old system?

The new text clarifies things. It's not encouraging anyone to use it, just stating the truth clearly.

Patch-set: 2