Update patch set 5
Patch Set 5:

R=go1.10

This causes object files to differ (unimportantly) between -c=1 and -c=2. Though that is in principle OK, it'd be better if it didn't, because the emphasis now is on correctness, and enforcing identical output is a useful way to accomplish that.

There are at least two ways this causes differing output.

(1) Handling of disable_checknil. That is issue 20242, and has an independently worthwhile fix in CL 42591.

(2) Different type info ordering. This is obvious in retrospect, because with c > 1, a second call to dumptypestructs is required. This can be fixed, but not without altering the behavior of -c=1, which we do not want to do for Go 1.9. Reproducing this is easy--just compile text/template with different values of c.

So let's defer this CL to Go 1.10. It only makes an noticeable difference for unusual/generated packages.

Patch-set: 5
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>