Update patch set 3
Patch Set 3:

Repeating in-person conversation:

ServeMux has never tried to play the performance game. There are dozens of ServeMux competitors out there, competing on features and speed. We've never tried to play that game. (Are we now?)

So the two main purposes of this code are 1) to be correct & stable, and 2) to be simple and idiomatic as a teaching aid. I'm a little afraid that the next round of Go learners reading ServeMux while writing their own ServeMuxes will interpret this code as idiomatic and infest their code with the sync/atomic package and friends.

So, I'm lukewarm at best on this. I'm not against it, but I'd like a second opinion from others.

Patch-set: 3
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>