Update patch set 1
Patch Set 1:

Options: 
golang.org/cl/39510
golang.org/cl/39511
golang.org/cl/39512
did not get much feedback and a decision within the last month. So here is a concrete cl to reduce the code as suggested in golang.org/cl/37959.

It would be nice if we can get this corrected in go1.9 to not need to ship a performance regression in go1.10 when slowing back Fields.

Also when we have found a good balance of generality, code size and performance for strings i can port these changes over to bytes to fix the performance differences for Fields between the packages.

Patch-set: 1
Reviewer: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>