Update patch set 1
Patch Set 1:

Adding this CL after 1.9 causes a regression from 1.9 to 1.10 since this CL slows down Fields for non-ascii. 

There is no regression in performance between 1.9 and 1.10 if we apply this CL for 1.9. as it is not slower than 1.8 but reduces the code size as was suggested in CL that speeds up fields for 1.9.

There is no regression in 1.9 without this CL but
There is one if we add this CL for 1.10 to bring down the code size again.

Patch-set: 1
Reviewer: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>