Update patch set 3
Patch Set 3:

This will probably end in the freeze.
So what would be best practice for me in mailing such a CL, without allocating human resources. So how to state that this should be for 1.10.
I often see R=go1.10 in a comment, but don't know the mechanism behind. Or should I make use of DO NOT REVIEW/SUBMIT ?
Sry for noise.

Patch-set: 3
Reviewer: Gerrit User 9850 <9850@62eb7196-b449-3ce5-99f1-c037f21e1705>