Update patch set 4
Patch Set 4:

Some observations: Changing readOp to int8 changes the alignment of the bootstrap buffer away from a 8 byte alignment and could result in performance change. There is a 2% performance drop in BufferFullSmallReads on my Ivy Bridge Cpu, have not looked further yet if that can be directly attributed to the alignment change or is because some other change (e.g a loop alignment somewhere).

Because the new Buffer struct has size 104 bytes on amd64 and 84 on 386 before and after due to padding and the next size class are 80 and 96 bytes I am unsure if this change saves any memory for the data. Might still matter for code size.

Patch-set: 4
CC: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>