Update patch set 2
Patch Set 2: Code-Review-2

(Temporary -2 to block submission during discussion)

Where is the bug about this?

Which test does this fix? Does this need a new test?

Is the only purpose of this change for external users using the syscall package, or for the runtime/std itself?

Is this fixed in x/sys/unix?

Patch-set: 2
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2