runtime: fix race on work.ndone
In runtime.gchelper, the compiler currently reorders the load of
work.ndone to after the atomic add of work.nwait (it shouldn't; this
is issue #20335). This creates a race since work.ndone can change as
soon as work.nwait is equal to work.ndone. If this happens, two or
more gchelpers can attempt to wake up the work.alldone note, cause a
"double wakeup" panic.

Work around this issue by making the load of work.nproc atomic. This
forces it to be loaded before the atomic add, fixing the race.

We should fix the underlying compiler issue, since it may be causing
similar issues in other places, but I think it's reasonable to make
this code less subtle independent of the compiler fix.

Fixes #19305 (I think).

Change-Id: Ieb1a84e1e5044c33ac612c8a5ab6297e7db4c57d