Update patch set 2
Patch Set 2:

OK I think what's happened here is I misunderstood the syscall vs x/sys/unix situation. All the relevent fixes done by this CL seem to have already been made in x/sys/unix (or, in the case of dirent, are not needed).

Is it better if this is just abandoned then given that regenerating ztypes in the syscall package is going to be so rare anyway?

Patch-set: 2
Reviewer: Gerrit User 17679 <17679@62eb7196-b449-3ce5-99f1-c037f21e1705>