Update patch set 2
Patch Set 2:

> Where is the bug about this?

There isn't one yet, I'll go and file one.

> Which test does this fix? Does this need a new test?

There don't appear to be any ptrace tests in x/sys/unix so it would need a new one. At the moment I have been testing it with a simple program which execs ls and reads the initial PC to see if it's sane (where "sane" depends on architecture and probably kernel config). At the moment on mips, PC always = 0 so for this specific instance we could check that PC != 0 after starting an executable.

> Is this fixed in x/sys/unix?

As I said in the other CL I was oblivious to the syscall and x/sys/unix situation. The bug doesn't seem to be fixed there. Should it be fixed in both packages or just in x/sys/unix?

Patch-set: 2
Reviewer: Gerrit User 17679 <17679@62eb7196-b449-3ce5-99f1-c037f21e1705>