Update patch set 19
Patch Set 19: Code-Review+1

I spoke with someone, they affirmed that BIND and QUERY statements will treat OUT and INOUT parameters differently. As such we include another field in the Out struct called "In" that when true will notify the driver (if applicable) that it is expecting to be passed to an INOUT parameter.

Patch-set: 19
Reviewer: Gerrit User 5080 <5080@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1